# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management API",
        "description": "Documentació oficial de l'API d'Inventari - Projecte Final",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/rental-contracts": {
            "get": {
                "tags": [
                    "Rental Contracts"
                ],
                "summary": "Llistar contractes de lloguer",
                "operationId": "841fd125d8a20d900e4fa7cefab75944",
                "parameters": [
                    {
                        "name": "provider_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_from",
                        "in": "query",
                        "description": "Data inici des de (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_to",
                        "in": "query",
                        "description": "Data fi fins a (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Llista de contractes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RentalContract"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rental Contracts"
                ],
                "summary": "Crear un nou contracte",
                "operationId": "249b422fccb6f74244ff2b3b2d1377c4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RentalContract"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creat correctament"
                    },
                    "422": {
                        "description": "Error de validació"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/rental-contracts/{id}": {
            "get": {
                "tags": [
                    "Rental Contracts"
                ],
                "summary": "Detalls d'un contracte",
                "operationId": "d668c003d65445ebbe421ff0ad9e4781",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No trobat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Rental Contracts"
                ],
                "summary": "Actualitzar contracte",
                "operationId": "571684f480a67d3da6b8bab63f925c9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "inactive"
                                    },
                                    "monthly_cost": {
                                        "type": "number",
                                        "example": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualitzat correctament"
                    },
                    "404": {
                        "description": "No trobat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rental Contracts"
                ],
                "summary": "Esborrar contracte",
                "operationId": "6514bef060394910a32febe023b08e74",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Esborrat correctament"
                    },
                    "404": {
                        "description": "No trobat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RentalContract": {
                "title": "RentalContract",
                "description": "Rental Contract model definition",
                "required": [
                    "provider_id",
                    "contract_number",
                    "start_date",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "provider_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "contract_number": {
                        "type": "string",
                        "example": "RC-2024-001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Main Office Lease",
                        "nullable": true
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2027-01-01",
                        "nullable": true
                    },
                    "monthly_cost": {
                        "type": "number",
                        "format": "float",
                        "example": 450.5,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "terms": {
                        "type": "string",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}